+++ test/Spec.hs
@@ -9,6 +9,7 @@ import qualified Canon.Antlr4.RoundTripTest
 import qualified Canon.Antlr4.VendoredTest
 import qualified Canon.AttachTest
 import qualified Canon.CanonicalCommentTest
+import qualified Canon.DecisionsTest
 import qualified Canon.Extract.Antlr4Test
 import qualified Canon.Model.CheckTest
 import qualified Canon.Git.ParseTest
@@ -37,6 +38,7 @@ main =
             , Canon.AttachTest.tests
             , Canon.Model.CheckTest.tests
             , Canon.Antlr4.InterpretTest.tests
+            , Canon.DecisionsTest.tests
             ]
         ]
     )